+++ java_console/trigger-ui/src/main/java/com/rusefi/trigger/TriggerImage.java
@@ -21,8 +21,10 @@ import java.util.List;
 /**
  * This utility produces images of trigger signals supported by rusEFI
  * <p>
+ *  rusefi_documentation\Images\triggers is the ultimate destination for these files
  * 06/23/15
  * Andrey Belomutskiy, (c) 2013-2020
+ * See TriggerWheelInfoTest for custom format
  */
 public class TriggerImage {
     private static final String OUTPUT_FOLDER = "triggers";
@@ -68,8 +70,6 @@ public class TriggerImage {
                 return "GM 24x 5";
             case TT_DODGE_NEON_1995:
                 return "Dodge Neon 1995";
-            case TT_DODGE_NEON_1995_ONLY_CRANK:
-                return "Dodge Neon 1995 crank only";
             case TT_SKODA_FAVORIT:
                 return "Skoda Favorit";
             case TT_GM_7X:
@@ -92,8 +92,8 @@ public class TriggerImage {
                 return "36/1";
             case TT_TOOTHED_WHEEL_36_2:
                 return "36/2";
-            case TT_TRI_TACH:
-                return "TriTach";
+// too dead            case TT_TRI_TACH:
+//                return "TriTach";
             case TT_TOOTHED_WHEEL_60_2:
                 return "60/2";
             case TT_GM_60_2_2_2:
@@ -102,7 +102,10 @@ public class TriggerImage {
         return triggerName.getTriggerName();
     }
 
-    public static void main(String[] args) throws InvocationTargetException, InterruptedException {
+    public static void main(String[] args) throws Exception {
+        String workingFolder = args.length > 0 ? args[0] : TriggerWheelInfo.DEFAULT_WORK_FOLDER;
+        TriggerMarkdownGenerator.generate(workingFolder, "triggers.md");
+
         SwingUtilities2.invokeAndWait(() -> {
             try {
                 runAwt(args);
@@ -160,7 +163,7 @@ public class TriggerImage {
             TriggerWheelInfo.readWheels(workingFolder, wheelInfo -> onWheel(triggerPanel, topPanel, content, wheelInfo));
         });
         Thread.sleep(1000L * sleepAtEnd);
-        System.exit(-1);
+        System.exit(0);
     }
 
     private static void onWheel(TriggerPanel triggerPanel, JPanel topPanel, JPanel content, TriggerWheelInfo triggerWheelInfo) {
@@ -400,7 +403,7 @@ public class TriggerImage {
 
             int h = getHeight();
 
-            g.drawString(name, 50, (int) (h * 0.75));
+            g.drawString(name, 50, (int) (h - f.getSize() * 3));
             if (id != null)
                 g.drawString(id, 0, (int) (h * 0.9));